Reimplemented the ability to override package suite actually intalled on the remote server.
New mechanism uses only single variable, which is undefined by default. In case it is undefined, appropriate default value for 'production','development' or 'demo' level server is filled in. In case it is defined and has correct value, it is used to enforce the package suite regardles in the server level classification.