Inverted logic in cleanup tasks to be disabled by default.
Previously cleanup of old and deprecated files was done by default. This was useful when doing migration from old version to current one. This task however considerably slows down the installation process on clean installations. The logic is now inverted, so the user has to explicitly request the cleanup operation.