Fixed the execution of mentat-controller.py and mentat-dbmngr.py as Nagios plugins.
When executed as Nagios plugins both scripts run as 'nagios' user. Therefore they must run in 'shell' mode so that they do not attempt to write to log file (to which they do not have permission to write). The mentat-dbmngr.py must not attempt to change to euid/egid Mentat.